fix: migrate networking client packet distributor and Recipe interface (major shape change)
- ClientPacketDistributor moved to net.neoforged.neoforge.client.network
- ServerPlayer#getServer() -> level().getServer()
- ViewportEvent#getPartialTick() now returns double, cast to float where needed
- Level.random -> getRandom() (DriftSmokeParticle)
- Recipe<T> interface reshaped: assemble()/getResultItem() dropped the
  HolderLookup.Provider param (assemble is back to single-arg, getResultItem()
  removed from the interface entirely -- kept as our own plain method);
  canCraftInDimensions() removed from the interface; added newly-required
  showNotification(), group(), placementInfo() (NOT_PLACEABLE, not grid-based),
  recipeBookCategory() (CRAFTING_MISC, never actually shown -- custom station UI)
- getSerializer()/getType() need the proper Recipe<ContainerRecipeInput> generic
  bound, not raw wildcards
- ItemStack#getItemHolder() -> getItem().getRegisteredName() directly

v0.0.0-beta.1

+++ src/main/java/com/skd/vehiclery/neoforge/VehicleryClientNeoForge.java
@@ -62,7 +62,7 @@ public class VehicleryClientNeoForge {
         });
 
         NeoForge.EVENT_BUS.<ViewportEvent.ComputeFov>addListener(evt ->
-                evt.setFOV((float) VehicleryClient.modifyBoostFov(Minecraft.getInstance(), evt.getFOV(), evt.getPartialTick())));
+                evt.setFOV((float) VehicleryClient.modifyBoostFov(Minecraft.getInstance(), evt.getFOV(), (float) evt.getPartialTick())));
     }
 
     @SubscribeEvent
